Fix: DS-CombiningDS-Lab ShoppingCenter

+++ 17.DS-CombiningDS-Lab/01.ShoppingCenter/01.ShoppingCenter/IShoppingCenter.cs
@@ -1,10 +1,10 @@
 ﻿public interface IShoppingCenter
 {
-    void AddProduct(string name, decimal price, string producer);
+    void AddProduct(string name, double price, string producer);
     void DeleteProducts(string producer);
     void DeleteProducts(string name, string producer);
     void FindProductsByName(string name);
     void FindProductsByProducer(string producer);
-    void FindProductsByPriceRange(decimal fromPrice, decimal toPrice);
+    void FindProductsByPriceRange(double fromPrice, double toPrice);
 }
